Rename ldbc.zio.telemetry -> ldbc.ziotelemetry package
